requirements.txt
streamlit
streamlit_lottie
PyMuPDF
Pillow